Add max prop to name in forms and refactor few of the schemas conneted to them
