--- src/forms/editPlacementSchema.ts
@@ -1,23 +1,0 @@
-import * as Yup from "yup";
-
-import { PLAYA_HEIGHT, PLAYA_VENUE_SIZE, PLAYA_WIDTH } from "settings";
-
-import { PlacementInput, VenueInput } from "api/admin";
-
-const initialMapIconPlacement: VenueInput["placement"] = {
-  x: (PLAYA_WIDTH - PLAYA_VENUE_SIZE) / 2,
-  y: (PLAYA_HEIGHT - PLAYA_VENUE_SIZE) / 2,
-};
-
-export const editPlacementSchema = Yup.object().shape<PlacementInput>({
-  addressText: Yup.string(),
-  notes: Yup.string(),
-  width: Yup.number().required("Required"),
-  height: Yup.number().required("Required"),
-  placement: Yup.object()
-    .shape({
-      x: Yup.number().required("Required").min(0).max(PLAYA_WIDTH),
-      y: Yup.number().required("Required").min(0).max(PLAYA_HEIGHT),
-    })
-    .default(initialMapIconPlacement),
-});
