+++ src/pages/Admin/Details/Form/DetailsForm.tsx
@@ -8,6 +8,7 @@ import classNames from "classnames";
 
 import {
   ALWAYS_EMPTY_ARRAY,
+  COMMON_NAME_MAX_CHAR_COUNT,
   DEFAULT_USER_STATUS,
   DEFAULT_VENUE_LOGO,
 } from "settings";
@@ -19,7 +20,7 @@ import { VenueTemplate } from "types/venues";
 
 import { adminWorldSpacesUrl } from "utils/url";
 
-import { venueV2Schema } from "forms/venueV2Schema";
+import { detailsSchema } from "forms/detailsSchema";
 
 import { useOwnedVenues } from "hooks/useConnectOwnedVenues";
 import { useUser } from "hooks/useUser";
@@ -142,7 +143,7 @@ const DetailsForm: React.FC<DetailsFormProps> = ({ venue, worldId }) => {
   } = useForm({
     mode: "onSubmit",
     reValidateMode: "onSubmit",
-    validationSchema: venueV2Schema,
+    validationSchema: detailsSchema,
     validationContext: {
       editing: !!venueId,
     },
@@ -376,7 +377,7 @@ const DetailsForm: React.FC<DetailsFormProps> = ({ venue, worldId }) => {
       </div>
       <AdminSpacesListItem title="The basics" isOpened>
         <>
-          <AdminSection title="Rename your space" withLabel>
+          <AdminSection title="Name your space" withLabel>
             <AdminInput
               name="name"
               placeholder="Space Name"
@@ -384,6 +385,7 @@ const DetailsForm: React.FC<DetailsFormProps> = ({ venue, worldId }) => {
               errors={errors}
               required
               disabled={nameDisabled}
+              max={COMMON_NAME_MAX_CHAR_COUNT}
             />
           </AdminSection>
           <AdminSection title="Subtitle" withLabel>
