+++ src/settings/playaSettings.ts
@@ -1,0 +1,31 @@
+import { CSSProperties } from "react";
+
+export const PLAYA_HEIGHT = 2000;
+export const PLAYA_WIDTH = 3000;
+export const PLAYA_AVATAR_SIZE = 48;
+export const PLAYA_VENUE_SIZE = 40;
+export const PLAYA_ICON_SIDE_PERCENTAGE = 5;
+
+// Burning Seed: playa is named paddock
+export const PLAYA_IMAGE = "/maps/paddock2k.jpg";
+
+export const PLAYA_VENUE_STYLES: Record<string, CSSProperties> = {
+  iconImage: {
+    width: PLAYA_VENUE_SIZE,
+    height: PLAYA_VENUE_SIZE,
+    overflow: "hidden",
+    borderRadius: "25%",
+    background: "rgba(147, 124, 99, 0.2)",
+    border: "2px solid rgba(147, 124, 99, 0.2)",
+    animation: "ripple 4s linear infinite",
+  },
+  draggableIconImage: {
+    width: PLAYA_VENUE_SIZE * 1.5,
+    height: PLAYA_VENUE_SIZE * 1.5,
+    overflow: "hidden",
+    borderRadius: "25%",
+    background: "rgba(147, 124, 99, 0.2)",
+    border: "2px solid rgba(147, 124, 99, 0.2)",
+    animation: "ripple 4s linear infinite",
+  },
+};
